Switch to the SCons build system, and perform major refactoring.
The kernel and drivers now live in nxos/base. Base produces a set of object
files that are useless by themselves, unless combined with one of the
"application kernels" in nxos/systems. Currently, a single system exists,
called 'marvin' (temporarily).

Aside from this huge refactoring and build system change, functionality should
be mostly the same (that includes a crash on shutdown in marvin, a bug I
introduced in an earlier change).