MOVEHUB: copy isr table from flash to SRAM
in order to use interrupts, the ISR vector table must be at 0x000000. Either flash (0x08000000) or SRAM (0x20000000) can be mapped to 0x00000000. Since the firmware starts at 0x8005000, we have to copy the table to SRAM and then tell the CPU to map SRAM to 0x00000000.

Also fixed up the bootloader magic value to use linker script instead of hard-coding address.