pbio: create new main module
This serves as the main entry point  for the library (well, initalization point anyway).

All of the init, deinit and poll functions are migrated here from MOVEHUB. (EV3 needs a bit of work to do the same).

This also adds an option to disable deinit functions to save space.

MOVEHUB fw size -76