MOVEHUB: add Makefile option to build and run a .mpy file from firmware
This adds a new option, PYBRICKS_MPY_MAIN_MODULE, to the MOVEHUB Makefile to allow specifying a .py file to compile to .mpy and append to the firmware.bin file

The has a few downsides, such as it makes the firmware nearly 1K larger and the .mpy file is copied from flash to RAM before running. This can be improved by modifying micropython to allow add a new option for running a raw (not frozen)  .mpy file from static memory (flash memory in our case).

The advatage to using this method is that users don't have to download the ARM toolchain to create frozen bytecode.

This option is disabled for now.

fw size +884