MOVEHUB: implement stop on button press
This adds a new pbsys feature to pass a stop function to pbsys_prepare_user_program(). This function will be called when the stop button is pressed while a user program is running.

In the MOVEHUB implementation, this is used to raise a SystemExit exception and perform a soft reset.

also fixed a couple global variables that should be static

fw size: +60