MOVEHUB: change stop button exception from SystemExit to KeyboardInterrupt
This is probably more intuitive to new(ish) users (in my experience, most people know about CTRL-C, but not CTRL-D).

This is also a bit better since SystemExit causes the REPL to reset, which could be suprising to unsuspecting users.

Finally, this also fixes a lock up when the stop button is pressed when no code is running.

Oh, and there was an error in the mp_hal_set_interrupt_char() declaration in mphalport.h that is corrected.

fw size -4