pbdrv: movehub: uart: add new UART driver
This adds a new UART driver for reading and writng characters to a UART device

A new callback is added to pbsys to provide a hook so that we can get a keyboard interrupt in MicroPython while code is running.

This is currently used to provide stdin/stout in MicroPython. However, this will eventually be made configurable so that port C is free for I/O devices.

Known issue: paste mode doesn't work in MicroPython because we can't process characters faster than they are being received.  Can try using a lower baud rate or software flow control (xon/xoff) to get around this.

fw size +76