sys/clock: use Contiki clock API instead of pbdrv
This is needed to integrate with other Contiki code. Fortunatly, the API is basically the same. We just had to add clock_usecs() since it is needed by MicroPython and is used in pbio already.

Also took the opportunity to move SysTickConfig() to the new clock_init() function.

MOVEHUB fw size -16