bricks/ev3dev: adjust imports
This brings the previous commit back to working state, with user scripts unchanged.

But instead of using

from boot import *

users would now do instead

from pybricks import *

That is, the user experience is the same but pybricks is now a package. This facilitates making a fake counterpart of the library.