extmod: parameters in c-module
This is a first step in reorganizing the modules at the C level, to make them more consistent with recently introduced package structure (#70)

This only organizes functionality between modules. To continue and define the package at the C level, we'd first need to add EV3 functionality to pbio, but this is not currently a priority.

User scripts remain unchanged by this commit.

Changes to pbio:
- button enum (numbers same as 1-9 keypad)
- color enum (numbers same as EV3)

movehub fw size: +468