bricks/ev3dev: improved color sensor value scaling based on logged data
sensor.reflection() # Same range as EV3-G, but 3x resolution: so 0.0 to 100.0
sensor.rgb() # Scaled to make it the same as above (0.0 to 100.0 for r, g, b) rather than arbitrary scaling

The `r` in rgb is now consistent with the value in reflection mode. Likewise, `blue` and `green` are normalized such that for a white surface, they give approximately the same value as `red`.  This compensation is especially useful for blue, which appears to have a much less bright LED.