modmotor: quick hack to make run work in tight loop for #89
This needs to be addressed more thoroughly in the next motor driver update, concerning the logic of starting (in main), running (in poll loop), and stopping maneuvers.

At the moment, when the user starts a maneuver, a status flag is set to "starting" and the control loop initializes the controller the first time it runs. Here it gets set to "running".  But this never happens if the user calls the command again before the control loop gets a chance to run; it's always "starting".

The real solution is to do control initialization outside the control loop.