bricks/ev3dev: add thread lock on motors
We need to protect against concurrent calls of pbio apis. Apparently, the unix port of Micropython has the GIL disabled for theads and we currently end up with a deadlock if we try to enable it. So we implment our own thread lock for now. Ideally, we will get rid of threads and use cooperative multitasking via Contiki threads instead.

Issue #101