travis: require matching micropython tag when we tag this repo
this requires the micropython repository to have a tag that matches the pattern "pybricks$TRAVIS_TAG"  when we tag the pybricks submodule. This way, it enforces us to make sure that we use the matching commit from micropython instead of the latest master in case we ever need to rebuild a tag.