pbio: move_hub: get BOOST interactive motor working
This adds some special cases for the BOOST interactive motor. In order to get speed and position updates, we have to send a special message (not sure of the meaning of the bytes at this point).

This still needs work if we want to use this motor as a "sensor", but speed control is working now.

fw size +392

Fixes #65