HUB4: Initial firmware that boots and lights up the LED
This is nearly a copy of the MOVEHUB drv source. A lot of code can probably be shared instead. At pbdrv, but also in bricks/ and Make

The pinout to the onboard devices are nearly identical compared to MOVEHUB

UART devices and motors not working yet. Pins are suspected to be the same (suspected that Port C&D on MOVEHUB = Port A&B on HUB4).

Checksum is slightly different due to different file size and different unused space at the end (4K).