bricks/ev3dev: simplify pbdrv and support dc motor
Instead of looping over the incremental motor numbers and reading port names from files, look for the device folder by using the port name directly.

This is nowhere near as good as #137, but it's less complicated than before. This makes it easier to do more checking and add support for DC motors.

This prepares support for DC motors at pbio level which ultimately extends to the PF2 system.