pbdrv/debug: add drivers for debug board
This adds drivers for a Nucleo-144 dev board used for debug. It has a button and separate red/green/blue LEDs, so it looks kind of like a powered up hub.

Serial console is using pins labeled tx and rx on the board.

LEDs, button and serial console are tested working.

Also including empty C files for other drivers so that we can share a makefile with other platforms.