bricks: rename brickconfig.h to mpconfigport.h
Due to the way upstream makefiles work, mpconfigport.h has to be in the same directory as the Makefile. So in order to move the common file to another directory, it has to have a different name.

So, change the order of the includes so that the platform-specific config file is name mpconfigport.h. The common file is renamed to stm32configport.h so that it has the same prefix as stm32.mk.