extmod/modlight: update light API
instead of

movehub.light(color)

we now do

movehub.light.color(color)
movehub.light.off()
movehub.light.brightness(red, green, blue) # Link to pbio not implemented yet

Because the light is an object (of the fictitious ColorLight class in the fake API). The movehub has one instance of such a light. This will enable us to support hubs with more than one light.

On ev3dev, the old syntax will co-exist with the new syntax for now. This keeps it backwards-compatible with the 1.0.0 release.

movehub fw size +164