src/uartdev: fix data not being read continuously
When the pbio_uartdev_receive_data() protothread ended, it could be up to 100ms before it was called again depending on what other events were going on in the system.

By adding a while loop, we start the next read as soon as one is finished to avoid this possible delay.

fixes 738c46ff
movehub fw size +12