drv/counter/ev3dev_stretch_iio: new driver
This is a new driver for the PRU-based quadrature encoders found in ev3dev-stretch. Instead of using /sys/class/tacho-motor/ devices that can disappear and reappear, we just use the IIO devices directly.

Also using udev to find the syfs path since it is not guaranteed to stay constant.

In the future, we could possibly improve performance by using IIO buffers instead of reading sysfs.