micropython-tag
This adds a new micropython-tag file that indicates the tag in the
parent micropython repository that is required to be checked out.

This way, we can rebase micropython without worring about breaking
older versions of pybricks-micropython. It will also notify developers
when changes have been made and they need to checkout a different
commit/tag.

New workflow should be to tag the micropython repository and update
the micropython-tag file each time a change is made in the micropython
repository.