extmod/pbkwarg: auto-parse and instantiate args
This introduces macros to considerably simplify the process of checking and parsing user arguments.

Each argument can be specified as required, or as optional with a default value of a desired type.

The macros also instantiate the arguments. The result is that arguments can be treated as if dealing with a standard MicroPython-style function of the form STATIC mp_obj_t  func(mp_obj_t a, mp_obj_t b), without having to worry about the right indexing of the *kw_args and union types to retrieve a and b.

Example:
```c
PB_PARSE_ARGS_FUNCTION(n_args, pos_args, kw_args,
    PB_ARG_REQUIRED(color),
    PB_ARG_DEFAULT_INT(brightness, 100)
);
```

Suppose the desired  arguments are x and y. The user is now allowed to do:
func(x, y)
func(x)
func(color=x)
func(color=x, brightness=y)
func(brightness=y, color=x)

And within the c-function, `color` and `brightness` are available as mp_obj_t, with already the correct values.