drv/adc/stm32_hal: add support for STM32L4
STM32L4 has a slightly different structure that F4. We can drop some paramaters that aren't being used since they will be initalized to the default value anyway.

And we have to add a #ifdef for one field name that is different.

Finally, ExternalTrigConv should be configurable since it depends on the timer that is selected.