drv/ev3dev_stretch/motor: use ioport driver to get syspath
This fixes the incorrect assumption that the port%d portion of the path has a deterministic value for %d. For example if input port 1 is disabled for kernel debugging, then it is only 3 + the port index instead of 4.

The motor init had to be moved after autostart process in order for ports to be available during motor init.