servo: get log by index
Instead of getting the whole data set as a Python object.

In practice, the user will be accessing the data in a loop, such as when writing to a file.

This avoids duplicating large datasets to Python user space. Duplicating can always be done at user level if desired.

When no index is given, the latest sample is returned. This allows the user to log data at specific times or events, such as triggering a sensor.