logger: extract out of servo
A generic logger keeps data and returns data to the user.

All the servo does is fill it with (servo-specific) data at each time stamp.

This will allow us to use the same logger for a drivebase, etc.

A few servo-specific bits still remain at this point, to be removed in subsequent commits.