bricks/nxt: change layout in linker script
- instead of using mapped address for RAM, use actual address of 2M
- move .text section first so that .vector table is read only

for some reason, since the vector table was in the .data section, the
linker was giving the wrong address for the IRQ handler functions (i.e.
it thought the functions were in RAM instead of ROM). This caused the
firmware to not start after flashing. By putting the vector table in the
.text section, things magically work correctly.