pbdrv/nxtcolor: simplify modes
Now it is just a sensor that measures everything, or a lamp. There is no reason to have different modes for each light. This makes use of a "payload" with the mode, inspired by lpf2 sensors.

Instead, the operations can now be controlled independently, so this change improves rather than reduces possibilities.

Also, use existing color definitions instead of making up new ones