modev3devices/AnalogSensor add resistance method
This assumes that the ADC measures the voltage across a resistive load which is in series with an internal resistor. The internal resistance was experimentally determined to be 10K using 8 known resistors. This seems to match the NXT and EV3 schematics. This gives reasonably accurate results, but they are slightly off at low resistance loads.