ev3dev/modules: Drop Python display/sound in hubs
The EV3Brick class will get new screen and speaker instances and/or modules written in C.

We will keep the old Python uev3dev implementation around for use in the compatibility module (pybricks.ev3brick) only. This way we can guarantee that the old API works as before, without imposing constraints on the new code.