extmod/modbuiltins/Speaker: add wait kwarg
For beep, say, and play, add a keyword argument wait=True which specifies whether the program should wait until the sound completes before continuing with the program.

This also adds a busy() method, which will return True if a sound is playing and False if not. This will allow a user to check/wait until a sound finishes playing.