modev3devices: add GyroSensor.calibrate()
This adds a non-standard functionality to reset sensor drift.

Now you can be sure that the sensor is not drifting, without having to unplug the cable. The operation still takes about 5 seconds, but it is much easier to keep the sensor stationary while calibrating.

This functionality initially was dropped from the Python implementation of the 1.0 release because it was unreliable, but it works well with the new C-implementation.