tools/mpybytes: in standalone, print C array
When used as a standalone script, __main__ will print the result in two formats that can be helpful to analyze or debug scripts.

In particularly, printing the C array will let us easily experiment with scripts on platforms that do not have a REPL or any form of I/O, like the NXT port.