stm32/main: add get message
This adds a function to wait for data of known length coming over the serial link, up to a given timeout.

The hub sends back a simple checksum after a message or after every chunk of data for larger messages.

Control resides with the sender (the IDE). If something is wrong, it can just stop sending and the hub will time out.