lib/ble5stack: implement missing functions
This implements quite a few functions we will need that were missing
from the TI code.

Also, it appears that the bt5stack we copied was for a different TI chip
other than the CC2640 we are using, so some functions had to be
modified to match what the real chip is expecting.

Some header file changes were copied from other versions of the TI
SimpleLink SDKs.