src/uartdev: remove check for unexpected mode
It is possible that a data message was already in progress when the mode
is changed, which could lead to an error condition. Since this is a
normal occurance, don't cause an error, but rather only update the
data received if the mode is the expected mode. If the mode of the
data messages does not change, a timeout will occur later anyway
because data->data_rec is not being updated.