drv/ioport/ioport_lpf2: store gpio value in dcm_data_t
This changes from using a local variable for storing the gpio value to
using a field in dcm_data_t. Since poll_dcm() is a proto thread, this
protects against loosing the value across PT_YIELD.

Although this is not a problem now, it will future-proof the code
against changes (like the one suggested in a TODO comment).

movehub fw size +0