src/uartdev: implement motor-specific max rate
This finishes a TODO that read the motor-specific max rate from the
motor and uses it to scale the speed percentage to get counts per
second.

A fallback of 1500 deg/sec is given for motors that don't provide this,
like the BOOST interactive motor.

Tests are updated with a test message of 100% speed so that we can
verify the max rate for each type of motor.

Also update count to non-zero value in tests to ensure we aren't just
getting the default value but rather what was actually read.

movehub fw size +84