pbio/main: reduce motor poll frequency
We are using 10 ms loops on the ev3dev port which is fast enough. On LPF2 ports, this is especially the case because new information doesn't come from the motor as fast.

This has the benefit of freeing up the CPU for other resources, and also reducing the amount of ram needed for data logging.