bricks/ev3dev/pbthread: replace locks with GIL
This replaces the pbthread locks with the MicroPython GIL. Previously,
we were running without the GIL enabled because there were blocking
functions in the pbio library. Hopefully all of these have been removed
now so that we can use the GIL since it greatly reduces the number of
locks needed.