extmod/modpupdevices: reenable light
This was temporarily disabled during the implementation of pbdevice.

Also, for off, just set it to SPEC1. This is not truly off, but dim enough compared to the full brightness lights.

We are intentionally using PROX/REFLT/AMBI to generate green, red, blue instead of using the setter in SPEC1 mode, to make the light bright enough to be considered a proper light.
 Because our code runs locally, the time needed to switch to these modes negligible anyway.