extmod/modpupdevices: add ir remote method
The ColorDistanceSensor can act as a PF1.0 infrared remote. This method adds only full on/off/direction control, by letting the user mimic remote presses.

This could probably not be enabled on movehub due to size considerations. It might be better to move it to a dedicated class, including speed levels, which can be excluded from movehub altogether.