extmod/modiodevices: define Ev3devSensor on ev3dev
This just has a different QSTR pointing to the LUMPDevice class. Since this was actually really acting as an Ev3devSensor on the ev3dev port, this is a better name for it. This will allow us to make it diverge from LUMPDevice in the future, by adding ev3dev-specific functionality.