extmod/modmotor: set dc mode for ev3dev
This is a hack to allows us to begin testing the mode switch, which is the same as already used by the sensors.

This is temporarily added at the module level, because at the moment we do not check for motor ID during init (and thus don't know if a mode switch is required).

But this will work the same for testing purposes; the device/port check just happens elsewhere in the code.