bricks/ev3dev/mpconfigport: disable MICROPY_ASYNC_KBD_INTR
This disables the MICROPY_ASYNC_KBD_INTR config option. When this option is enabled, the exception is raised from the unix signal callback, which means it could happen anywhere in C code, including on the background contiki thread and when the GIL is not held. In other words, there are lots of opportunities for bad timing.

By disabling this option, the KeyboardInterrupt exception will be scheduled for later via the mp_handle_pendiing() function. This way the exception is only thrown between VM instructions or in the event poll hook callback. This way, we can properly write functions to handle being interrupted since we know excactly when it could happen.