bricks/ev3dev/image: always fill background in print() method
This modifies the print() method so that the background is always
filled. This makes the text easier to read if there was something
already in the background. It also fixes the bug where the foreground
was the last color set by the draw_text() method. Now, the foreground
color is always black.