pbio/trajectory: implement trajectory patching
This implements the patch function, but it is not used by the servo yet.

This allows a time based maneuver to be patched onto an existing one if it is tangent to the original. By using an existing exact starting point, we avoid having to pause and restart the rate integrator, avoiding roundoff errors.

This most trivial example (a simplified special case) is where a user keeps calling the same maneuver in a tight loop. In this case, the trajectory would not be updated each time and the PID integration state would not be reset. This makes the behavior as smooth as if it were called just once.