bricks/ev3dev/mpconfigport: disable GIL VM divisor
This sets MICROPY_PY_THREAD_GIL_VM_DIVISOR to disable it. We can
have some long-running functions, like motor_Motor_run that can cause
the background thread to be blocked for too long if the divisor is
enabled.

This is a performace hit since we are now constantly releasing and
acquireing the GIL mutex lock, but it is necessary for motor control.
It might be posible to set the divisor to a low number, such as 4 to
get a few % back on performace at the risk of affecting motor control.
So we are playing it safe for now and disabling it completly.