bricks/ev3dev/brickconfig: remove g_main_context_iteration from MICROPY_VM_HOOK_LOOP
This is a performance improvement. In an unscientific test, this doubles
the number of micropython instructions processed per second.

This means that the glib loop will now only be run during
MICROPY_EVENT_POLL_HOOK. This should be OK since we are only using it
for long running things like implementing mp_hal_delay_ms and playing
sounds.