extmod/modmotor: add internal busy flag
This will not be exposed because it does not belong here, hence the underscore.

It should instead be a method of the control object, and it should call a pbio function that evaluates is_done_func for that controller. However, this is more work and we need it right now for testing.