extmod/modev3devices: don't use GYRO-G&A mode
The G&A mode seems broken on ev3dev-stretch. Using a 19N3 sensor,
changing the mode to G&A causes the sensor to get stuck in that mode
and it cannot be changed back to any other mode.

It also makes a kernel worker thread jump to ~18% when the robot is
perfectly still. In contrast, the RATE mode makes the kworker thread
about 6 or 7 percent when the robot is still and it spiked as high as
14% when the robot was shaken.