bricks/ev3dev: stop motors and sound on unhandled exception
When using the REPL, e.g. using the pause button in the ev3dev VS Code
extension, we want the robot to stop when it is interrupted.

Motors and beep are (hopefully) the only physical outputs that could be
running in the background that can cause problems (motors can break
stuff and endless beep is really annoying).