lib/ev3dev/lego_port: use status to read port mode
If the status is OK, we do not need to switch port mode.

This reduces the need to switch modes in case of nxt analog sensors. We have previously removed the user option to choose whether to verify type. Instead, it is now always set to analog mode if the status is not analog already.

This also prevents a mode switch for LUMPDevice if a device is auto-detected as a lego-sensor already.