src/math: fix pbio_math_div_i32_fix16()
pbio_math_div_i32_fix16() was truncating the dividend and the quotient
to a fix16_t which caused incorrect results when the dividend is larger
than INT16_MAX or the divisor is less than one.

This is probably not the most efficient or precise implementation, but
it is a quick hack to expand the range of inputs for the function.