pbio/drivebase: adopt control settings from servo
Because the drivebase operates on the sum and difference of the counts of both motors, the control settings carry over quite easily.

For example, if the max speed of one motor is 1000 counts per second, then the maximum drivebase counts per second is 2000, because its count state is just two motor counts added.

By doing this for all settings, the drivebases perform approximately as well as their single motor counterparts. This way, we don't have to code different defaults for different platforms (EV3/Powered Up), because this is already reflected in the respective servo settings for each platform.