pbio/servo: relax tolerances and control gains
Reduce the gain on the medium motor, for better performance in a wider speed and torque application range.

Relax the tolerances which govern when the "done" flag is raised. This makes most motion sequences faster, because it doesn't wait for one motion to precisely hit the target before doing the next command.