pbio/motorpoll: use stop functions on reset
Also use PBIO_ERROR_AGAIN, not PBIO_SUCCESS, as a non-zero indicator that we should poll. This way, we don't poll all motors until we are told to start polling it.

Also ensure not to set error status on reset all. This way, motors do stop but keep polling, so they can be used even after an unhandled exception.