tools/runserial, mpybytes: full rewrite
This updates and cleans up the tools for use with latest firmware, including verification of checksums on chunks of data.

It also drops the dependency on mpy_cross from PyPI. Instead the user is asked to provide a path to the mpy_cross executable, such as the one built in this development repository.

Other improvements is checking the subprocess status for failure so we capture syntax error or missing file errors.

[skip ci]