bricks/ev3dev/modules: rename boot.py to core.py
We are not using it in the same was as MicroPython "boot.py", so we rename it to avoid confusion.

It is merely a shortcut to import core functionality quickly in the REPL; it does not run automatically on boot.

Also, make the imports match the imports of the new project template. This covers core functionality, which are common classes/functions that cover most examples and typical robot projects that use one EV3 kit.