pbio/control: integrate speed on sign change
Only let the speed integrator pause if it is stuck in the reference direction. Otherwise it would not resume if the speed reference suddenly changed into the other direction, which can occur if you quickly change a remote control dial, for example.

Tightening this condition is not required for position based control, since there we have no ambiguity over the sign of proportional control, which is always the same as the position error.