extmod/modmotor,modrobotics: change default stop to HOLD
This is done to be consistent with existing LEGO programming platforms.
Also, most control transitions will behave better when using HOLD as
opposed to COAST.

An exception is made for "run until stalled" since this would usually
result in trying to hold a high torque position. It still uses COAST
by default.