bricks/ev3dev/Image: fix SEGV in load_image
This fixes a segmentation fault in the load_image() method when using
the special "_screen_" object.

Since load_image_buffer_obj was static just a pointer and not a
structure, the allocated memory would eventually get garbage collected
since it was not pointed to by any root pointers. This fixes the problem
by adding a field to store the extra object instead of using a static
variable.