bricks/stm32: always include main.mpy
Previously, there was a compile option to either
include a main.mpy file in the user flash, or include
download-and-run functionality. This change enables
both.

- The main program is started/stopped with the button.
- Running programs from the IDE is unchanged.

Now it is possible to simply insert a new main.mpy without
installing and configuring the whole toolchain. This can
be done by the IDE, to let end-users create a dedicated
firmware file with their own main.py built in.