bricks/ev3dev/Image: Use MP memory allocation for image data
This changes the image class to use MicroPython's memory allocator to
allocate the framebuffer memory for images instead of letting the GRX3
library allocate the memory.

This applies more pressure on the garbage collector to trigger a
collection when an Image object is no longer referenced. Without this,
MicroPython would never collect Image objects and we would get into
swap memory and cause the EV3 to become sluggish and have to be
rebooted.