pbdrv/counter/stm32f0: fix IRQ handler
This fixes problems that arise when two Move Hub motors rotate
at the same time. Introduced in:
a38301e17f676bf35a6e0aca077afe1cee98929b

Since then, the interrupt handler updated both motor counts for a
single interrupt. This reverts the check for which motor it was.

The check is hard coded because IRQs are not supported in
the gpio driver yet.