bricks/ev3dev/pbdevice: delay after mode switch
When changing to a new mode, some sensors continue to send
data that was recorded when that mode was used last, even if that
was a long time ago. If that mode was not used since power on,
the stale data is usually zero.

To ensure we only give new data to the user, we add a short delay
after switch to a new mode. The required delay depends on the
selected sensor and mode, and theoretically also the mode that
we were in before.

This is especially critical for sensors like the Color Sensor, which is
often used for various purposes (and hence different modes) in FLL
robots. Stale data could make robots behave different than intended
and different each time depending on previous conditions.

A delay like this is also present in the original EV3 software. For
the Color Sensor, this was measured to be approximately 30 milliseconds.

This change also allows to remove a few hardcoded delays elsewhere
for some NXT devices, in subsequent commits.