pbio/control: fix relative travel by 0 angle
Previously, this behavior was not well-defined in case of trajectory patching. For now, this enforces hold behavior, going vertically to the origin in the angle/speed phase plot. This matches the behavior of angles > 0

In the future, relative angles >= 0 with speed in excess of what it can decelerate in time without overshoot, should curve around the origin of the phase plot, then decelerate smoothly back to the target. This is already done for angles < 0.