extmod/modrobotics: bind settings by limits
The settings() just set lower defaults for straight() and turn(). When
they call pbio functions, they are still bound by the control limits.

By limiting the values kept as class attributes, this implicit bound is
made explicit because it makes it possible to see that very large
values are capped.