bricks/stm32: generate firmware metadata and zip
This adds build targets to generate a firmware.metadata.json file that
contains information about the firmware.bin file. This is intended to
be used by firmware flashing programs to match the firmware to the
correct type of hub and to allow replacing the user .mpy file without
having to use a linker.