bricks/stm32: include main.py in firmware.zip
This changes the firmware.zip file format to include an incomplete base
firmware binary blob and a main.py file. It is up to the flasher program
to compile main.py to main.mpy, append it to the firmware-base.bin and
add a checksum. Additional metadata is included to provide the flasher
program what it needs to know in order to accomplish this.