drv/cplus_hub/bluetooth: only send write response if not command
BlueZ will cause a device disconnect because of a protocol error if we
don't do the right thing. Also it is more efficient if we do the right
thing.

There are two types of GATT characteristic write: with response
(request) and without response (command). As the names indicate, we
should not send a response for the latter.