drv/uart/uart_stm32l4_ll: new STM32L4 UART driver
This adds a new UART driver tht uses the low-level API from the STM32
library. This allows us to use DMA instead of interrupts on the UARTs.

This driver is enabled on Control+. This fixes UART devices reconnecting
because of missed Rx data. It also fixes port D not working because of
baud rate issues.