bricks/stm32: require import from pybricks package
Previously, we were automatically importing classes from some modules
to save space. However, this is too implicit and inconsistent with EV3
MicroPython. This brings in a MicroPython patch that fakes importing
from the Pybricks package. This allows the first users to have a
consistent experience until a real solution is implemented or until
we revert to frozen .py dummy files to provide the package structure.