platform/stm32: adjust interrupt priorities
On STM32F0 MCUs, the interrupt priority is only a 2-bit value, so 0 is
the most urgent and 3 is the least urgent (lower priority number will
preempt higher priority number).

On these (move hub and city hub), we make UART have priority of 0,
position/speed of built-in motors have priority 1 and everything else
priority 3. It is especially important to allow UART to preempt all
other interrupts so that we don't miss incoming data.

On STM32F4/L4, the HAL library configures interrupts for 16-bit priority
and 16-bit subpriority. Priority determines preemption as above and
subpriority determines precedence if two interrupts happen at the same
time. Again, UART used in interrupt mode needs to be able to preempt
all other interrupts. But UART in DMA mode it is not as urgent to handle
interrupts. Other priority values were chosen arbitrarily and could be
reconsidered in the future.