extmod/modpupdevices: Add InfraredSensor
Using the raw data, we can get much more information than the 10 distance increments that this sensor normally gives.

We provide two convenience methods based on this data:
- distance() nonlinearly scales the raw data to provide an approximately
  linear distance value. It corresponds to the original sensor value,
  but with higher resolution and a scale from 0-100 to match similar
  relative distance methods on the other sensors. Resolution is somewhat
  course above 50%.
- reflection() linearly scales the data to provide reflection data close
  to the sensor. This can be used for line following or precise distance
  sensing up close.

There is also a count() method, which does the same as the original
sensor. In this class, the count is reset to 0 upon instantiation. Using
the other methods does not reset this count.