@pybricks/firmware: download assets instead of building
This modifies the @pybricks/firmware npm package to download the
firmware zip file from GitHub instead of building them locally. This
way we have fewer firmware versions floating around.

The npm package also now has the firmware version in the build
metadata portion of the the package version. This is enforced by
having the firmware download script parse package.json for this
version.