extmod/modpupdevices: drop SPEC1 for reflection
This follows the same logic as the previous change for distance mode. However, iIt turns out that (R+G+B)/12 equals the reflection value. By using the RGB mode we avoid a mode switch between color and reflection.

We could also get better resolution by returning a float, but for now we keep an integer result to save memory. The maximum is currently a bit over 100 as well since we don't discard high values.