github: update to ARM Embedded GCC 9
GCC 9 is fairly stable now. Also this is first version that is signed
and notarized for macOS making it the only version that runs on macOS.

Previous builds of the toolchain were built with -O0, so this version
can compile twice as fast and also reduces firmware size a bit.