drv/pwm: new driver subsystem
This introduces an abstraction layer for PWM devices. Currently this
just consists of STM32 TIM devices being used as PWMs. It prepares the
way for adding other types of PWM devices.

The platform-specific motor and light drivers are converted to use the
new APIs. This is one step towards making those drivers more generic.