drv/light: use math instead of lookup table
This replaces lookup tables with linear equations to get reasonable
colors from the LEDs. The numbers were derived from the SunLED
XZM2CRKM2DGFBB45SCCB datasheet. This LED looks like the one in the
Move hub and the City hub - not sure if it is an exact match though.
This also produces reasonable colors on the Control+ hub as well.