pybricks: move version to package, drop custom uos
It is better to have the fully fledged version of moduos.c only on hubs
that can support it, instead of having various different and reduced
implementations of it across all the hubs.

Since pybricks now becomes more like a Python package, it is fairly
natural to provide the version in the package root instead. It has
a similar tuple format as `uos.uname()` used to have, but we can
adapt this format to be more consistent with Python packages.

This also reduces movehub firmware size by 80 bytes since it removes
both the module and the function that were wrapped around the
version tuple.