drv/led: new LED drivers
This consolidates the platform-specific light drivers into a couple of
new LED drivers.

Instead of color lookup tables, drivers can now show the full spectrum
of hues in the HSV color model plus white.

Currently, the same calibration values are used on all platforms. The
values were derived from the Move/City hub LEDs so the colors are a bit
off on the Control+ hub since it has a different LED.

All LED PWM periods are now set to 10000 to match the Control+ hub.
This way we don't have to have platform data for different periods on
different devices. (Control+ hub shares LED PWMs with motor PWMs, so
it can't be changed.)