drv/pwm_tlc5955_stm32: enable enhanced PWM
This enables the ES-PWM feature on the TI TLC5955 chip. This feature
breaks up the PWM signal into smaller pulses during each PWM period
of 65535 counts. This helps reduce flicker when the LEDs are filmed.

GSCLK is currently set to 9.6 MHz and the chip uses 65536 counts for
the PWM period, so the PWM rate is only 146.5 Hz.