drv/usb/stm32_usb_serial: fix lockup when USB disconnected
When USB is disconnected while transmitting, the usual transmit complete
callback is not called which prevented any future transmissions. Also
the MicroPython implementation of print() would block forever waiting
for more room in the transmit buffer when USB was disconnected.

Fix by implementing some connection state logic and resetting the Tx and
Rx busy state when USB is connected.